+++ test/test_UPF_MAP/main.cpp
@@ -47,9 +47,9 @@ public:
     { }
 
 protected:
-    bool runCondition() override
+    bool run_condition() override
     {
-        if (getFilteringStep() < simulation_steps_)
+        if (step_number() < simulation_steps_)
             return true;
         else
             return false;
@@ -65,11 +65,11 @@ protected:
         return sis_filenames;
     }
 
-    bool initialization() override
+    bool initialization_step() override
     {
         estimates_extraction_.setMethod(EstimatesExtraction::ExtractionMethod::map);
 
-        if (!SIS::initialization())
+        if (!SIS::initialization_step())
             return false;
 
         /* Initialize initial mean and covariance for each particle. */
@@ -88,9 +88,9 @@ protected:
         return true;
     }
 
-    void filteringStep() override
+    void filtering_step() override
     {
-        SIS::filteringStep();
+        SIS::filtering_step();
 
         /* Update previous corrected particles. */
         previous_corr_particle_ = cor_particle_;
@@ -102,7 +102,7 @@ protected:
 
         for (std::size_t i = 0; i < probabilities.rows(); i++)
             for (std::size_t j = 0; j < probabilities.rows(); j++)
-                probabilities(i, j) = prediction_->getStateModel().getTransitionProbability(previous_corr_particle.col(j), corr_particle.col(i)).coeff(0);
+                probabilities(i, j) = prediction().getStateModel().getTransitionProbability(previous_corr_particle.col(j), corr_particle.col(i)).coeff(0);
 
         return probabilities;
     }
@@ -113,7 +113,7 @@ protected:
 
         /* Get likelihoods. */
         VectorXd likelihoods;
-        std::tie(std::ignore, likelihoods) = correction_->getLikelihood();
+        std::tie(std::ignore, likelihoods) = correction().getLikelihood();
 
         /* Get transition probability matrix. */
         MatrixXd probabilities = getTransitionProbabilityMatrix(cor_particle_.state(), previous_corr_particle_.state());